+++ CHANGELOG.rst
@@ -1,7 +1,8 @@
 1.0.5 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix bytes serialization issue
+  [vangheem]
 
 
 1.0.4 (2017-05-01)